requirements.txt
py-cord==2.0.0b4
psycopg2~=2.9.3
SQLAlchemy~=1.4.31
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pygsheets~=2.0.5